+++ etils/epath/gpath.py
@@ -110,11 +110,6 @@ class _GPath(abstract_path.Path):
   def _backend(self) -> backend_lib.Backend:
     try:
       backend = _PREFIX_TO_BACKEND[self._uri_scheme]
-      # Choose tf_backend if tf is installed. We don't use FSSpec by default
-      # for retro-compatibility, because needed dependencies (gcsfs or s3fs)
-      # may not be installed. fsspec_backend was indeed introduced later.
-      if _is_tf_installed() and self._uri_scheme is not None:
-        return backend_lib.tf_backend
       return backend
     except KeyError:
       supported = ', '.join(f'`{k}://`' for k in _PREFIX_TO_BACKEND)